POC: Implement basic endpoint for fetching workflow executions from containerized Temporal (#6)
This is a very naive implementation of an endpoint that can talk to the containerized version of Temporal to fetch a list of workflow executions.

The types need work and there are clearly more query parameters that we'll need to support. There are edge cases left unhandled. But, it's a quick proof of concept that will allow us to work against the fixtures in the `samples-go` repository as we build out some of the pages.

To-do:

- Make this work with single-binary, web-v2 server.
- Figure out how we're going to bundle this with the single binary (we may choose not to and only use these endpoints in the cloud version).
- Add support for the rest of the query parameters currently supported.
- Add better handling.
- Refine types.
- Add support for respecting next page tokens.
- (Probably much more)