Add proposal for Box and Text
I don't feel super strongly about either of these proposals, but they're mostly meant to get the conversation started.

The `Text` component just renders a string, but takes care of the translation. The cool thing about it is that it doesn't render the slot. So, you _could_ just put a string of text in there like "Recent Workflows" just like we had before we did internationalization.

The `Box` component is an abstraction around the native DOM. It does support the same types as the DOM element that it wraps, but it gives us an inflection point where we could do stuff like dark mode logic or adding a standardized format for `data-test-id` or whatever.

The idea is that we could use this as the basis for the common set of things we use in the UI (e.g. `Input`, `Heading`, etc.) and have a place to handle accessibility or theming logic for us.