Update base images
The base-builder and base-server images are ~4months old at this point
and the security scanner I'm using picks up a CVE-2023-45288.

This CVE is resolved in go >= 1.22.2 so is likely already resolved
in base-builder 1.14.11 since that was built recently and 1.22.5
is the latest in golang:1.22-alpine3.20 if someone pulls it today.